add AuditingRepositoryDecoratorFactory
